# Crosscheck with Paper, complex matrix optimizations