--- a/src/MonteCarlo.jl
+++ b/src/MonteCarlo.jl
@@ -35,7 +35,7 @@ include("helpers.jl")
 export enable_benchmarks, disable_benchmarks, print_timer, reset_timer!
 include("linalg/general.jl")
 include("linalg/UDT.jl")
-# include("linalg/complex.jl") # TODO
+include("linalg/complex.jl") # TODO
 include("linalg/blockdiagonal.jl")
 include("flavors/abstract.jl")
 include("models/abstract.jl")