--- a/src/MonteCarlo.jl
+++ b/src/MonteCarlo.jl
@@ -76,7 +76,7 @@ export neighbors, directions
 include("flavors/MC/MC.jl")
 include("flavors/DQMC/main.jl")
 export Greens, GreensAt, CombinedGreensIterator
-export GreensMatrix, dagger
+export GreensMatrix, swapop
 export boson_energy_measurement, greens_measurement, occupation, magnetization
 export charge_density, charge_density_correlation, charge_density_susceptibility
 export spin_density, spin_density_correlation, spin_density_susceptibility