--- a/src/configurations.jl
+++ b/src/configurations.jl
@@ -91,6 +91,7 @@ See also [`ConfigRecorder`](@ref), [`Discarder`](@ref)
 """
 mutable struct BufferedConfigRecorder{CT} <: AbstractRecorder
     filename::FilePath
+    link_id::String
     buffer::Vector{CT}
     rate::Int64
     idx::Int64
@@ -98,11 +99,12 @@ mutable struct BufferedConfigRecorder{CT} <: AbstractRecorder
     total_length::Int64
     save_idx::Int64
 end
-function BufferedConfigRecorder(fn::String, buffer, rate, idx, chunk, N, sidx)
-    BufferedConfigRecorder(AbsolutePath(fn), buffer, rate, idx, chunk, N, sidx)
+function BufferedConfigRecorder(fn::String, link_id, buffer, rate, idx, chunk, N, sidx)
+    BufferedConfigRecorder(AbsolutePath(fn), link_id, buffer, rate, idx, chunk, N, sidx)
 end
 function BufferedConfigRecorder{CT}(filename, rate = 10, chunk_size = 1000) where CT
-    BufferedConfigRecorder(filename, Vector{CT}(undef, chunk_size), rate, 1, 1, 0, -1)
+    link_id = string(rand(UInt128))
+    BufferedConfigRecorder(filename, link_id, Vector{CT}(undef, chunk_size), rate, 1, 1, 0, -1)
 end
 function BufferedConfigRecorder(MC::Type, M::Type, filename; rate = 10, chunk_size = 1000)
     BufferedConfigRecorder{compressed_conf_type(MC, M)}(filename, rate, chunk_size)
@@ -195,31 +197,49 @@ function load_chunk!(cr::BufferedConfigRecorder, chunk)
 end
 
 function update_filepath!(cr::BufferedConfigRecorder, parent_path)
+    # this only messes with relative filepaths
+    # - update absolute path if necessary (parent path changed)
+    # - move and rename on collision if link_id doesn't match
+    # - replace if link_id matches
+
     if cr.filename.is_relative
-        # get dir of savefile
+        # get dir of parent savefile
         path, _ = splitdir(parent_path)
 
         # get adjusted absolute path, removing potential preceding /
         rp = cr.filename.relative_path
         filepath = joinpath(path, startswith(rp, '/') ? rp[2:end] : rp)
 
-        # move file if the path changed
+        # move file if the absolute path is incorrect
         if filepath != cr.filename.absolute_path
-            # TODO
-            # add a random tag to the cr file and the dqmc savefile on save
-            # if tag matches we can replace, else rename
-            
-            # source exists, move while avoiding overwriting (?)
             if isfile(cr.filename.absolute_path)
-                if isfile(filepath) 
-                    new_filepath = _generate_unique_filename(filepath)
-                    @warn "Config file '$filepath' already exists, renaming to '$new_filepath' to avoid collision."
-                    filepath = new_filepath
+                target_link_id = try 
+                    JLD2.jldopen(filepath, "r") do f
+                        get(f, "link_id", "N/A") # maybe default to current?
+                    end
+                catch e # if we can't read the link_id we do not want to touch the file
+                    "ERROR"
                 end
-                mv(cr.filename.absolute_path, filepath)
+
+                # file exists and does not match this id - rename our file
+                if target_link_id != cr.link_id
+                    # rename
+                    new_filepath = _generate_unique_filename(filepath)
+                    @warn(
+                        "There already exists an independent file at " * 
+                        "$filepath. Renaming to $new_filepath to avoid " *
+                        "overwriting other data."
+                    )
+                    filepath = new_filepath
+
+                # file exists and has this id - replace it
+                else
+                    # replace
+                    @warn "Replacing dublicate file at $filepath"
+                end
+
+                mv(cr.filename.absolute_path, filepath, force = true)
             end
-            # if source does not exist we keep the filename. If that file already
-            # exists we assume it's this simulations file.
         end
 
         # update filepath
@@ -231,6 +251,13 @@ end
 
 
 function _save(file, cr::BufferedConfigRecorder, entryname::String="configs")
+    # save link_id
+    JLD2.jldopen(cr.filename.absolute_path, "a+") do file
+        if !haskey(file, "link_id") 
+            file["link_id"] = cr.link_id
+        end
+    end
+
     # save current buffer
     if !(cr.chunk == -1 && cr.idx == -1) # not in unitialized state
         if cr.save_idx != cr.total_length || !isfile(cr.filename.absolute_path)
@@ -242,9 +269,10 @@ function _save(file, cr::BufferedConfigRecorder, entryname::String="configs")
     update_filepath!(cr, file.path)
 
     # main save information
-    write(file, entryname * "/VERSION", 2)
+    write(file, entryname * "/VERSION", 3)
     write(file, entryname * "/tag", "BufferedConfigRecorder")
     write(file, entryname * "/filename", cr.filename)
+    write(file, entryname * "/link_id", cr.link_id)
     write(file, entryname * "/buffer", cr.buffer)
     write(file, entryname * "/rate", cr.rate)
     write(file, entryname * "/total_length", cr.total_length)
@@ -253,16 +281,28 @@ function _save(file, cr::BufferedConfigRecorder, entryname::String="configs")
 end
 
 function _load(data, ::Val{:BufferedConfigRecorder})
-    if !(data["VERSION"] <= 2)
+    if !(data["VERSION"] <= 3)
         throw(ErrorException("Failed to load BufferedConfigRecorder version $(data["VERSION"])"))
     end
+
+    link_id = get(data, "link_id", "N/A")
+    
     cr = BufferedConfigRecorder(
-        data["filename"], data["buffer"], data["rate"], -1, -1, 
+        data["filename"], link_id, data["buffer"], data["rate"], -1, -1, 
         data["total_length"], data["save_idx"]
     )
 
     # adjust relative FilePath
     update_filepath!(cr, data.path)
+
+    # generate new link_id if none exists yet
+    if link_id == "N/A"
+        link_id = rand(UInt128)
+        cr.link_id = link_id
+        JLD2.jldopen(cr.filename.absolute_path, "a+") do file
+            file["link_id"] = cr.link_id
+        end
+    end
 
     return cr
 end