--- a/src/flavors/DQMC/greens.jl
+++ b/src/flavors/DQMC/greens.jl
@@ -64,8 +64,11 @@ exponentials from left and right.
 Inplace version of `greens`.
 """
 @bm function greens!(mc::DQMC; output=mc.stack.greens_temp, input=mc.stack.greens, temp=mc.stack.Ur)
+    # TODO rework measurements to work well with StructArrays and remove this    
     if isdefined(mc.stack, :complex_greens_temp)
-        GreensMatrix(0, 0, _greens!(mc, mc.stack.complex_greens_temp, output, input, temp))
+        _greens!(mc, output, input, temp)
+        copyto!(mc.stack.complex_greens_temp, output)
+        GreensMatrix(0, 0, mc.stack.complex_greens_temp)
     else
         GreensMatrix(0, 0, _greens!(mc, output, input, temp))
     end
@@ -92,19 +95,6 @@ function _greens!(
     return target
 end
 
-# TODO rework measurements to work well with StructArrays and remove this
-function _greens!(
-        mc::DQMC_CBFalse, 
-        target = mc.stack.complex_greens_temp, temp_target = mc.stack.greens_temp, 
-        source = mc.stack.greens, temp = mc.stack.Ur
-    )
-    eThalfminus = mc.stack.hopping_matrix_exp
-    eThalfplus = mc.stack.hopping_matrix_exp_inv
-    vmul!(temp, source, eThalfminus)
-    vmul!(temp_target, eThalfplus, temp)
-    copyto!(target, temp_target)
-    return target
-end
 
 function _greens!(
         mc::DQMC_CBTrue, target::AbstractMatrix = mc.stack.greens_temp,