--- a/src/flavors/DQMC/main.jl
+++ b/src/flavors/DQMC/main.jl
@@ -135,6 +135,8 @@ include("greens.jl")
 include("measurements/generic.jl")
 # Quick constructers and measurement kernels (i.e. applied Wicks theorem)
 include("measurements/measurements.jl")
+# Superfluid stiffness stuff
+include("measurements/superfluid_stiffness.jl")
 # Contains some post processing tools
 include("measurements/extensions.jl")
 # structs and conversions from the old system