… docs/src/examples/ALF1.md
… docs/src/examples/HBC_model.jl
… docs/src/examples/HofmannBergChowdhury.md
… docs/src/examples/assets/HBC/CDS.png
… docs/src/examples/assets/HBC/DS.png
… docs/src/examples/assets/HBC/HBC_lattice.png
… docs/src/examples/assets/HBC/HoffmannBerg_CDS.png
… docs/src/examples/assets/HBC/HoffmannBerg_PS.png
… docs/src/examples/assets/HBC/HoffmannBerg_SDCz.png
… docs/src/examples/assets/HBC/PS.png
… docs/src/examples/assets/HBC/SDCz.png
… docs/src/examples/assets/fig1_comparison.png
… docs/src/examples/assets/fig2_comparison.png
… docs/src/examples/assets/fig3_comparison.png
… docs/src/examples/assets/fig6_comparison.png
… docs/src/examples/assets/fig7_comparison.png
… docs/src/examples/assets/triangular/fig1_comparison.png
… docs/src/examples/assets/triangular/fig2_comparison.png
… docs/src/examples/assets/triangular/fig3_comparison.png
… docs/src/examples/assets/triangular/fig6_comparison.png
… docs/src/examples/assets/triangular/fig7_comparison.png
… docs/src/examples/assets/triangular/tri_Hub_ref1_1.png
… docs/src/examples/assets/triangular/tri_Hub_ref1_2.png
… docs/src/examples/assets/triangular/tri_Hub_ref2_1.png
… docs/src/examples/assets/triangular/tri_Hub_ref2_2.png
… docs/src/examples/assets/triangular/tri_Hub_ref3_1.png
… docs/src/examples/assets/triangular/tri_Hub_ref3_2.png
… docs/src/examples/assets/triangular/tri_Hubbard_CDSF.png
… docs/src/examples/assets/triangular/tri_Hubbard_SDS.png
… docs/src/examples/triangular_Hubbard.jl