@ffreyer 2021-09-21
https://carstenbauer.github.io/MonteCarlo.jl/dev/examples/triangular_Hubbard/ clearly does not `\Delta^\dagger \Delta + \Delta |delta^\dagger` for its pairing correlation. It fits way worse than `\Delta \Delta^\dagger`. Needs to be adjusted if I go through with switching the default from `pc_kernel` to `pc_combined_kernel`.