@ffreyer 2021-09-27
Making use of StructArrays in measurements is tough atm. There is no left vs right multiplier like in `mul!`, matrix elements could technically be multiplied by `i` or conjugates could be taken. So instead I'm just converting complex `StructArrays` back to plain complex matrices. This isn't optimal, but still a lot better... (~2s -> ~1.33s)